Move defaults module to client-modules folder
